+++ src/main/java/subway/line/StationLineGroup.java
@@ -1,6 +1,6 @@
 package subway.line;
 
-import subway.subway.Station;
+import subway.station.Station;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
