+++ src/main/java/subway/station/StationController.java
@@ -1,4 +1,4 @@
-package subway.subway;
+package subway.station;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
