+++ src/main/java/subway/station/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.subway;
+package subway.station;
 
 import org.springframework.data.jpa.repository.JpaRepository;
