+++ src/main/java/subway/station/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.subway;
+package subway.station;
 
 public class StationRequest {
     private String name;
