+++ src/main/java/subway/station/StationService.java
@@ -1,4 +1,4 @@
-package subway.subway;
+package subway.station;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
